Create gudpy class and put iterators into single module
