--- gudpy/core/iterators/thickness.py
@@ -1,31 +1,0 @@
-from core.iterators.iterator import Iterator
-
-
-class ThicknessIterator(Iterator):
-    """
-    Class to represent a Thickness Iterator. Inherits Iterator.
-    This class is used for iteratively tweaking the thickness of a
-    flatplate sample.
-    This means running gudrun_dcs iteratively, and adjusting the thickness
-    of each sample across iterations. The new thickness is determined by
-    applying a coefficient calculated from the results of gudrun_dcs in the
-    previous iteration.
-    ...
-
-    Methods
-    ----------
-    applyCoefficientToAttribute
-        Multiplies a sample's thicknesses by a given coefficient.
-    organiseOutput
-        Organises the output of the iteration.
-    """
-
-    name = "IterateByThickness"
-
-    def applyCoefficientToAttribute(self, object, coefficient):
-        # Determine a new total thickness.
-        totalThickness = object.upstreamThickness + object.downstreamThickness
-        totalThickness *= coefficient
-        # Assign the new thicknesses.
-        object.downstreamThickness = totalThickness / 2
-        object.upstreamThickness = totalThickness / 2
